ecommerce_analytics/requirements.txt
apache-airflow-providers-snowflake==5.1.1
snowflake-connector-python==3.5.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
great-expectations==0.17.19
